+++ Sources/SentryCrash/Recording/SentryCrash.m
@@ -34,6 +34,7 @@
 #import "SentryCrashMonitor_AppState.h"
 #import "SentryCrashMonitor_System.h"
 #import "SentryCrashReportFields.h"
+#import "SentryCrashReportStore.h"
 #import "SentryCrashSystemCapabilities.h"
 #import <NSData+Sentry.h>
 
@@ -95,7 +96,6 @@ getBasePath()
 @synthesize userInfo = _userInfo;
 @synthesize deleteBehaviorAfterSendAll = _deleteBehaviorAfterSendAll;
 @synthesize monitoring = _monitoring;
-@synthesize deadlockWatchdogInterval = _deadlockWatchdogInterval;
 @synthesize onCrash = _onCrash;
 @synthesize bundleName = _bundleName;
 @synthesize basePath = _basePath;
@@ -179,12 +179,6 @@ getBasePath()
 - (void)setMonitoring:(SentryCrashMonitorType)monitoring
 {
     _monitoring = sentrycrash_setMonitoring(monitoring);
-}
-
-- (void)setDeadlockWatchdogInterval:(double)deadlockWatchdogInterval
-{
-    _deadlockWatchdogInterval = deadlockWatchdogInterval;
-    sentrycrash_setDeadlockWatchdogInterval(deadlockWatchdogInterval);
 }
 
 - (void)setOnCrash:(SentryCrashReportWriteCallback)onCrash
@@ -407,6 +401,28 @@ SYNTHESIZE_CRASH_STATE_PROPERTY(BOOL, crashedLastLaunch)
     return nil;
 }
 
+- (NSArray<NSString *> *)getScreenshotPaths:(int64_t)reportID
+{
+    char report_screenshot_path[SentryCrashCRS_MAX_PATH_LENGTH];
+    sentrycrashcrs_getScreenshotPath_forReportId(reportID, report_screenshot_path);
+    NSString *path = [NSString stringWithUTF8String:report_screenshot_path];
+
+    BOOL isDir = false;
+    if (![NSFileManager.defaultManager fileExistsAtPath:path isDirectory:&isDir] || !isDir)
+        return @[];
+
+    NSArray *files = [NSFileManager.defaultManager contentsOfDirectoryAtPath:path error:nil];
+    if (files == nil)
+        return @[];
+
+    NSMutableArray *result = [[NSMutableArray alloc] initWithCapacity:files.count];
+    [files enumerateObjectsUsingBlock:^(id _Nonnull obj, NSUInteger idx, BOOL *_Nonnull stop) {
+        [result addObject:[NSString stringWithFormat:@"%@/%@", path, obj]];
+    }];
+
+    return result;
+}
+
 - (void)doctorReport:(NSMutableDictionary *)report
 {
     NSMutableDictionary *crashReport = report[@SentryCrashField_Crash];
@@ -452,6 +468,7 @@ SYNTHESIZE_CRASH_STATE_PROPERTY(BOOL, crashedLastLaunch)
                              | SentryCrashJSONDecodeOptionIgnoreNullInObject
                              | SentryCrashJSONDecodeOptionKeepPartialObject
                                error:&error];
+
     if (error != nil) {
         SentryCrashLOG_ERROR(
             @"Encountered error loading crash report %" PRIx64 ": %@", reportID, error);
@@ -460,6 +477,12 @@ SYNTHESIZE_CRASH_STATE_PROPERTY(BOOL, crashedLastLaunch)
         SentryCrashLOG_ERROR(@"Could not load crash report");
         return nil;
     }
+
+    NSArray *screenShots = [self getScreenshotPaths:reportID];
+    if (screenShots.count > 0) {
+        crashReport[SENTRYCRASH_REPORT_SCREENSHOT_ITEM] = screenShots;
+    }
+
     [self doctorReport:crashReport];
 
     return crashReport;