+++ Sources/SentryCrash/Recording/SentryCrashC.c
@@ -30,7 +30,6 @@
 #include "SentryCrashFileUtils.h"
 #include "SentryCrashMonitorContext.h"
 #include "SentryCrashMonitor_AppState.h"
-#include "SentryCrashMonitor_Deadlock.h"
 #include "SentryCrashMonitor_System.h"
 #include "SentryCrashMonitor_User.h"
 #include "SentryCrashMonitor_Zombie.h"
@@ -61,6 +60,7 @@ static bool g_shouldPrintPreviousLog = false;
 static char g_consoleLogPath[SentryCrashFU_MAX_PATH_LENGTH];
 static SentryCrashMonitorType g_monitoring = SentryCrashMonitorTypeProductionSafeMinimal;
 static char g_lastCrashReportFilePath[SentryCrashFU_MAX_PATH_LENGTH];
+static void (*g_saveScreenShot)(const char *) = 0;
 
 // ============================================================================
 #pragma mark - Utility -
@@ -105,6 +105,18 @@ onCrash(struct SentryCrash_MonitorContext *monitorContext)
         sentrycrashcrs_getNextCrashReportPath(crashReportFilePath);
         strncpy(g_lastCrashReportFilePath, crashReportFilePath, sizeof(g_lastCrashReportFilePath));
         sentrycrashreport_writeStandardReport(monitorContext, crashReportFilePath);
+    }
+
+    // Report is saved to disk, now we try to take screenshots.
+    // Depending on the state of the crash this may not work
+    // because we gonna call into non async-signal safe code
+    // but since the app is already in a crash state we don't
+    // mind if this approach crashes.
+    if (g_saveScreenShot) {
+        char crashScreenshotsPath[SentryCrashCRS_MAX_PATH_LENGTH];
+        sentrycrashcrs_getScreenshotsPath_forReport(
+            g_lastCrashReportFilePath, crashScreenshotsPath);
+        g_saveScreenShot(crashScreenshotsPath);
     }
 }
 
@@ -168,14 +180,6 @@ sentrycrash_setUserInfoJSON(const char *const userInfoJSON)
 }
 
 void
-sentrycrash_setDeadlockWatchdogInterval(double deadlockWatchdogInterval)
-{
-#if SentryCrashCRASH_HAS_OBJC
-    sentrycrashcm_setDeadlockHandlerWatchdogInterval(deadlockWatchdogInterval);
-#endif
-}
-
-void
 sentrycrash_setIntrospectMemory(bool introspectMemory)
 {
     sentrycrashreport_setIntrospectMemory(introspectMemory);
@@ -209,6 +213,12 @@ void
 sentrycrash_setMaxReportCount(int maxReportCount)
 {
     sentrycrashcrs_setMaxReportCount(maxReportCount);
+}
+
+void
+sentrycrash_setSaveScreenshots(void (*callback)(const char *))
+{
+    g_saveScreenShot = callback;
 }
 
 void
@@ -297,4 +307,10 @@ void
 sentrycrash_deleteReportWithID(int64_t reportID)
 {
     sentrycrashcrs_deleteReportWithID(reportID);
+}
+
+bool
+sentrycrash_hasSaveScreenshotCallback()
+{
+    return g_saveScreenShot != NULL;
 }