+++ Tests/SentryTests/TestClient.swift
@@ -82,6 +82,7 @@ class TestClient: Client {
     var captureCrashEventInvocations = Invocations<(event: Event, scope: Scope)>()
     override func captureCrash(_ event: Event, with scope: Scope) -> SentryId {
         captureCrashEventInvocations.record((event, scope))
+        print("### Captured")
         return SentryId()
     }